+++ src/gcs.ts
@@ -27,7 +27,7 @@ import {
   type ObjectStream,
   type ParsedRange,
   type StoreErrorClassifiers,
-} from "./index.js";
+} from "./index.ts";
 
 // Re-export for convenience
 export { ObjectNotFoundError, ObjectChangedError, StoreUnavailableError };