+++ src/s3.ts
@@ -45,7 +45,7 @@ import {
   type ObjectStream,
   type ParsedRange,
   type StoreErrorClassifiers,
-} from "./index.js";
+} from "./index.ts";
 
 // Re-export for convenience
 export { ObjectNotFoundError, ObjectChangedError, StoreUnavailableError };